# [cli] add compile cache to improve startup performance