--- a/.changeset/stupid-seals-exercise.md
+++ b/.changeset/stupid-seals-exercise.md
@@ -1,0 +1,5 @@
+---
+"vercel": minor
+---
+
+[cli] add compile cache to improve startup performance