--- a/packages/cli/scripts/build.mjs
+++ b/packages/cli/scripts/build.mjs
@@ -49,7 +49,4 @@ copyFileSync(
   new URL('src/util/get-latest-version/get-latest-worker.js', repoRoot),
   new URL('get-latest-worker.js', distRoot)
 );
-copyFileSync(
-  new URL('src/vc.js', repoRoot),
-  new URL('vc.js', distRoot)
-);
+copyFileSync(new URL('src/vc.js', repoRoot), new URL('vc.js', distRoot));