This improves the startup of Vercel CLI from 18ms to 13ms on my MacBook Pro M3 running Node.js 22.

References

- https://nodejs.org/api/module.html#moduleenablecompilecachecachedir
- https://devblogs.microsoft.com/typescript/announcing-typescript-5-7/#support-for-v8-compile-caching-in-node.js