@styfle 2025-01-10
I just benchmarked it again running 4 times each and the difference is like 300ms on the first run and then around 60ms.

## Before
```
▲ time ./packages/cli/dist/index.js --version
0.23s user 0.09s system 49% cpu 0.646 total
▲ time ./packages/cli/dist/index.js --version
0.18s user 0.03s system 100% cpu 0.209 total
▲ time ./packages/cli/dist/index.js --version
0.18s user 0.02s system 102% cpu 0.195 total
▲ time ./packages/cli/dist/index.js --version
0.18s user 0.02s system 102% cpu 0.197 total
```

## After
```
▲ time ./packages/cli/dist/vc.js --version   
0.17s user 0.06s system 65% cpu 0.353 total
▲ time ./packages/cli/dist/vc.js --version   
0.12s user 0.03s system 100% cpu 0.148 total
▲ time ./packages/cli/dist/vc.js --version
0.12s user 0.02s system 103% cpu 0.135 total
▲ time ./packages/cli/dist/vc.js --version
0.12s user 0.02s system 102% cpu 0.136 total
```